fix(authentication-service): export all artifacts of modules from root
export models and providers from modules to
make them able to import from service by user

BREAKING CHANGE:
this will affect the projects where imports are from dist

gh-1950
