+++ services/authentication-service/src/index.ts
@@ -6,6 +6,7 @@ import './load-env';
 export * from './component';
 export * from './keys';
 export * from './models';
+export * from './modules/auth';
 export * from './providers';
 export * from './repositories';
 export * from './services';
