+++ services/authentication-service/src/modules/auth/controllers/index.ts
@@ -1,0 +1,11 @@
+export * from './apple-login.controller';
+export * from './azure-login.controller';
+export * from './cognito-login.controller';
+export * from './facebook-login.controller';
+export * from './google-login.controller';
+export * from './instagram-login.controller';
+export * from './keycloak-login.controller';
+export * from './login.controller';
+export * from './logout.controller';
+export * from './otp.controller';
+export * from './saml-login.controller';
